+++ src/components/app/App.js
@@ -22,6 +22,7 @@ import {AlertsProvider} from '../../containers/alerts'
 import {MarkdownTipsDialogProvider} from '../../containers/markdown-tips'
 import {FooterProvider} from '../../containers/footer'
 import {SoundsProvider} from '../../containers/sounds'
+import {ManageContactsDialogProvider} from '../../containers/manage-contacts'
 
 @injectSheet(styles)
 export default class App extends PureComponent {
@@ -59,6 +60,7 @@ export default class App extends PureComponent {
           <BillingWarningProvider />
           <ChannelMembersInviteProvider />
           <NewConversationProvider />
+          <ManageContactsDialogProvider />
           <InviteToOrgProvider />
           <NotificationSettingsProvider />
           <UnreadChannelsProvider />