+++ src/components/manage-contacts/index.js
@@ -1,0 +1,1 @@
+export ManageContactsDialog from './ManageContactsDialog'