+++ src/components/tabs/index.js
@@ -1,0 +1,2 @@
+export Tab from './Tab'
+export TabsNav from './TabsNav'