+++ src/legacy/redux-emitter/index.js
@@ -127,10 +127,6 @@ class ReduxEmitter extends Emitter {
     this.emit('triggerRoomManager')
   }
 
-  showPmManager() {
-    this.emit('triggerPMManager')
-  }
-
   toggleOrgSettings(elem) {
     this.emit('toggleOrgSettings', elem)
   }